Use 16 bits per channel instead of 32.

+++ chafa/chafa.h
@@ -26,7 +26,7 @@ ChafaColorSpace;
 /* Color space agnostic, using fixed point */
 typedef struct
 {
-  gint ch [4];
+  gint16 ch [4];
 }
 ChafaColor;
